+++ types/evidence.go
@@ -747,9 +747,9 @@ func (e PhantomValidatorEvidence) ValidateBasic() error {
 		return errors.New("empty vote")
 	}
 
-	// if err := e.Header.ValidateBasic(); err != nil {
-	// 	return fmt.Errorf("invalid header: %v", err)
-	// }
+	if err := e.Header.ValidateBasic(); err != nil {
+		return fmt.Errorf("invalid header: %v", err)
+	}
 
 	if err := e.Vote.ValidateBasic(); err != nil {
 		return fmt.Errorf("invalid signature: %v", err)
@@ -850,9 +850,9 @@ func (e LunaticValidatorEvidence) ValidateBasic() error {
 		return errors.New("empty vote")
 	}
 
-	// if err := e.Header.ValidateBasic(); err != nil {
-	// 	return fmt.Errorf("invalid header: %v", err)
-	// }
+	if err := e.Header.ValidateBasic(); err != nil {
+		return fmt.Errorf("invalid header: %v", err)
+	}
 
 	if err := e.Vote.ValidateBasic(); err != nil {
 		return fmt.Errorf("invalid signature: %v", err)
